Add the Player class
This commit adds the Player class which holds the player's info, in addition,
I added a method to the String class which helps to know if the content of
a given string is compound of numbers.

Another improvemen: In previous commits I added color in some text, I deleted
those colors because it could cause errors if someone doesn't have the 'colorize'
Gem installed.