Move the test of each class in their own file
This commit moves the test of each class in their own file and instead of comparing
the result of the method #board in the class Play now test the type of result. The
test of the method UserInterface#play was removed because there are multiple
I/O with are not inside the method rather calling private methods.